+++ src/collective/solr/solr.py
@@ -28,6 +28,7 @@
 # c.delete('123')
 # c.commit()
 
+from copy import deepcopy
 from plone.dexterity.utils import safe_unicode
 import six.moves.http_client
 import socket
@@ -42,6 +43,7 @@ from collective.solr.utils import getConfig
 from collective.solr.utils import translation_map
 
 from logging import getLogger
+import base64
 import six
 
 logger = getLogger(__name__)
@@ -55,6 +57,8 @@ class SolrConnection:
         persistent=True,
         postHeaders={},
         timeout=None,
+        login=None,
+        password=None,
     ):
         self.host = host
         self.solrBase = str(solrBase)
@@ -66,27 +70,42 @@ class SolrConnection:
         # a real connection to the server is not opened at this point.
         self.conn = six.moves.http_client.HTTPConnection(self.host, timeout=timeout)
         # self.conn.set_debuglevel(1000000)
+        self.login = login
+        self.auth_headers = {}
+        if login and password:
+            credentials = "{0}:{1}".format(login, password)
+            authorization = base64.b64encode(credentials.encode("ascii"))
+            self.auth_headers["Authorization"] = "Basic {0}".format(
+                authorization.decode("ascii")
+            )
         self.xmlbody = []
         self.xmlheaders = {"Content-Type": "text/xml; charset=utf-8"}
         self.xmlheaders.update(postHeaders)
+        self.xmlheaders.update(self.auth_headers)
         if not self.persistent:
             self.xmlheaders["Connection"] = "close"
         self.formheaders = {
-            "Content-Type": "application/x-www-form-urlencoded; charset=utf-8"
+            "Content-Type": "application/x-www-form-urlencoded; charset=utf-8",
         }
+        self.formheaders.update(self.auth_headers)
         if not self.persistent:
             self.formheaders["Connection"] = "close"
 
     def __str__(self):
+        xmlheaders = deepcopy(self.xmlheaders)
+        if "Authorization" in xmlheaders:
+            xmlheaders["Authorization"] = "Basic ***"
         return (
             "SolrConnection{host=%s, solrBase=%s, persistent=%s, "
-            "postHeaders=%s, reconnects=%s}"
+            "postHeaders=%s, reconnects=%s, login=%s, password=%s}"
             % (
                 self.host,
                 self.solrBase,
                 self.persistent,
-                self.xmlheaders,
+                xmlheaders,
                 self.reconnects,
+                self.login,
+                self.auth_headers and "***" or None,
             )
         )
 
@@ -336,7 +355,9 @@ class SolrConnection:
         # schema_url = '%s/admin/file/?file=managed-schema'
         logger.debug("getting schema from: %s", schema_url % self.solrBase)
         try:
-            self.conn.request("GET", schema_url % self.solrBase)
+            self.conn.request(
+                "GET", schema_url % self.solrBase, headers=self.auth_headers
+            )
             response = self.conn.getresponse()
         except (
             socket.error,
@@ -346,7 +367,9 @@ class SolrConnection:
         ):
             # see `doPost` method for more info about these exceptions
             self.__reconnect()
-            self.conn.request("GET", schema_url % self.solrBase)
+            self.conn.request(
+                "GET", schema_url % self.solrBase, headers=self.auth_headers
+            )
             response = self.conn.getresponse()
 
         if response.status == 200: